+++ backend/config.go
@@ -1,6 +1,9 @@
 package backend
 
 import (
+	"crypto/sha1"
+	"encoding/hex"
+	"fmt"
 	"log"
 	"os"
 	"path"
@@ -13,6 +16,7 @@ func GetUserConfigPath() string {
 	}
 	var dir = path.Join(userConfigDir, "Nothing Search")
 	createUserConfigPath(dir)
+	createUserConfigPath(path.Join(dir, "thumbnails"))
 	return dir
 }
 
@@ -25,4 +29,19 @@ func createUserConfigPath(dir string) {
 	if err := os.Mkdir(dir, os.ModePerm); err != nil {
 		log.Fatal("Creating user dir failed", err)
 	}
+}
+
+func GetThumbnailPath(filePath string) string {
+	configPath := GetUserConfigPath()
+	// extension := filepath.Ext(filePath)
+	encodedName := calculateSHA1(filePath)
+	name := fmt.Sprintf("%s.%s", encodedName, "jpg")
+	return path.Join(configPath, "thumbnails", name)
+}
+
+func calculateSHA1(input string) string {
+	hasher := sha1.New()
+	hasher.Write([]byte(input))
+	sha1Hash := hex.EncodeToString(hasher.Sum(nil))
+	return sha1Hash
 }